enhancement/APP-Update-auth-flow
- Update auth flow
- Update tokens models and generation flow
- Refactor according to taken code style

+++ controllers/auth.js
@@ -1,8 +1,9 @@
 // outsource dependencies
 import bcrypt from 'bcrypt';
 import jwt from 'jsonwebtoken';
+import { v4 as uuidv4 } from 'uuid';
+import { body } from 'express-validator';
 import { StatusCodes } from 'http-status-codes';
-import { body, validationResult } from 'express-validator';
 
 // local dependencies
 import UserModel from '../models/User.js';
@@ -11,21 +12,14 @@ import { TOKEN_TYPE } from '../constants/index.js';
 
 export const requireAuth = async (req, res, next) => {
   try {
-    const authHeader = req.header('Authorization');
-    if (!authHeader?.startsWith('Bearer ')) {
-      return res.status(StatusCodes.UNAUTHORIZED).json({ message: 'Authorization data is invalid', errorCode: 'unauthorized' });
-    }
-    const accessToken = authHeader.replace('Bearer ', '');
+    const accessToken = req.header('Authorization').replace('Bearer ', '');
 
     const decoded = jwt.verify(accessToken, process.env.JWT_ACCESS_SECRET);
-    const token = await TokenModel.findOne({ _id: decoded.tid, user: decoded.id, type: TOKEN_TYPE.ACCESS }).exec();
-    if (!token) {
-      return res.status(StatusCodes.UNAUTHORIZED).json({ message: 'Authorization data is invalid', errorCode: 'unauthorized' });
-    }
+    const token = await TokenModel.findOne({ session: decoded.sid, user: decoded.id, type: TOKEN_TYPE.ACCESS, value: accessToken }).exec();
+    if (!token) { throw new Error('unauthorized'); }
+
     // Attach authenticated user and Access Token to request object
-    req.userId = decoded.id;
-    req.token = accessToken;
-    req.tokenId = decoded.tid;
+    req.authData = { userId: decoded.id, token: accessToken, tokenId: decoded.sid };
     next();
   } catch (error) {
     return res.status(StatusCodes.UNAUTHORIZED).json({ message: 'Authorization data is invalid', errorCode: 'unauthorized' });
@@ -66,10 +60,6 @@ export const registerValidator = [
 
 export const register = async (req, res) => {
   try {
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-      return res.status(StatusCodes.UNPROCESSABLE_ENTITY).json({ errors: errors.array(), errorCode: 'validation_failed', message: errors.array()[0]?.msg });
-    }
     const { firstName, lastName, password, email, occupation } = req.body;
     const existingUser = await UserModel.findOne({ email }).exec();
     if (existingUser) {
@@ -106,44 +96,40 @@ const deleteUserExpiredTokens = async userId => {
   }
 };
 
-const generateAccessToken = async user => {
-  const accessTokenExp = Math.floor(Date.now() / 1e3) + Number(process.env.JWT_ACCESS_LIFETIME || 3600);
-  const accessTokenDoc = await new TokenModel({
-    user: user._id,
-    type: TOKEN_TYPE.ACCESS,
-    expires: new Date(accessTokenExp * 1e3),
-  }).save();
-  return jwt.sign({
-    id: user._id,
-    type: 'access',
-    email: user.email,
-    exp: accessTokenExp,
-    tid: accessTokenDoc._id,
-    lastName: user.lastName,
-    firstName: user.firstName,
-  }, process.env.JWT_ACCESS_SECRET);
-};
-
 export const login = async (req, res) => {
   try {
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-      return res.status(StatusCodes.UNPROCESSABLE_ENTITY).json({ errors: errors.array(), errorCode: 'validation_failed', message: errors.array()[0]?.msg });
-    }
     const { email, password } = req.body;
     const user = await UserModel.findOne({ email }).select('password').exec();
     if (!user) { return res.status(StatusCodes.NOT_FOUND).json({ message: 'Invalid credentials', errorCode: 'invalid_credentials' }); }
     const isPasswordValid = await bcrypt.compare(password, user.password);
     if (!isPasswordValid) { return res.status(StatusCodes.NOT_FOUND).json({ message: 'Invalid credentials', errorCode: 'invalid_credentials' }); }
-    const accessToken = await generateAccessToken(user);
-
-    const refreshTokenExp = Math.floor(Date.now() / 1e3) + Number(process.env.JWT_REFRESH_LIFETIME || 86400);
-    const refreshTokenDoc = await new TokenModel({
+    const sid = uuidv4();
+    const accessToken = jwt.sign({
+      sid,
+      id: user._id,
+      email: user.email,
+      type: TOKEN_TYPE.ACCESS,
+      lastName: user.lastName,
+      firstName: user.firstName,
+    }, process.env.JWT_ACCESS_SECRET, { expiresIn: Number(process.env.JWT_ACCESS_LIFETIME || 3600) });
+    const decodedAccessToken = jwt.decode(accessToken);
+    const accessTokenDoc = await new TokenModel({
+      session: sid,
       user: user._id,
-      type: TOKEN_TYPE.REFRESH,
-      expires: new Date(refreshTokenExp * 1e3),
+      value: accessToken,
+      type: TOKEN_TYPE.ACCESS,
+      expires: new Date(decodedAccessToken.exp * 1e3),
     }).save();
-    const refreshToken = jwt.sign({ tid: refreshTokenDoc._id, id: user._id, type: 'refresh', exp: refreshTokenExp }, process.env.JWT_REFRESH_SECRET);
+
+    const refreshToken = jwt.sign({ sid, id: user._id, type: 'refresh' }, process.env.JWT_REFRESH_SECRET, { expiresIn: Number(process.env.JWT_REFRESH_LIFETIME || 86400) });
+    const decodedRefreshToken = jwt.decode(refreshToken);
+    const refreshTokenDoc = await new TokenModel({
+      session: sid,
+      user: user._id,
+      value: refreshToken,
+      type: TOKEN_TYPE.REFRESH,
+      expires: new Date(decodedRefreshToken.exp * 1e3),
+    }).save();
 
     await deleteUserExpiredTokens(user._id);
 
@@ -162,16 +148,11 @@ export const logoutValidator = [
 
 export const logout = async (req, res) => {
   try {
-    const authHeader = req.header('Authorization');
-    if (!authHeader?.startsWith('Bearer ')) {
-      return res.status(StatusCodes.UNAUTHORIZED).json({ message: 'Authorization data is invalid', errorCode: 'unauthorized' });
-    }
-    const accessToken = authHeader.replace('Bearer ', '');
-
+    const accessToken = req.header('Authorization').replace('Bearer ', '');
     const decoded = jwt.verify(accessToken, process.env.JWT_ACCESS_SECRET, { ignoreExpiration: true });
     const { refreshToken } = req.body;
     const decodedRefresh = jwt.verify(refreshToken, process.env.JWT_REFRESH_SECRET, { ignoreExpiration: true });
-    await TokenModel.deleteMany({ $and: [{ user: decoded.id }], $or: [{ _id: decoded.tid }, { _id: decodedRefresh.tid }] }).exec();
+    await TokenModel.deleteMany({ $and: [{ user: decoded.id }], $or: [{ session: decodedRefresh.sid }] }).exec();
 
     await deleteUserExpiredTokens(decoded.id);
     return res.status(StatusCodes.OK).json({ success: true });
@@ -185,21 +166,34 @@ export const logout = async (req, res) => {
 
 export const refreshAccessToken = async (req, res) => {
   try {
-    const authHeader = req.header('Authorization');
-    if (!authHeader?.startsWith('Bearer ')) {
-      return res.status(StatusCodes.UNAUTHORIZED).json({ message: 'Authorization data is invalid', errorCode: 'unauthorized' });
-    }
-    const expiredAccessToken = authHeader.replace('Bearer ', '');
+    const expiredAccessToken = req.header('Authorization').replace('Bearer ', '');
     const decoded = jwt.verify(expiredAccessToken, process.env.JWT_ACCESS_SECRET, { ignoreExpiration: true });
     const { refreshToken } = req.body;
     try {
       const decodedRefresh = jwt.verify(refreshToken, process.env.JWT_REFRESH_SECRET);
-      if (decoded.id !== decodedRefresh.id) { throw new Error('invalid-refresh-or-access-token'); }
+      if ((decoded.id !== decodedRefresh.id) || (decoded.sid !== decodedRefresh.sid)) { throw new Error('invalid-refresh-or-access-token'); }
       await deleteUserExpiredTokens(decodedRefresh.id);
-      const isTokenExist = await TokenModel.exists({ _id: decodedRefresh.tid, user: decodedRefresh.id }).exec();
-      if (!isTokenExist) { throw new Error('invalid-refresh-token'); }
+      const refreshTokenDoc = await TokenModel.findOne({
+        sid: decodedRefresh.sid,
+        user: decodedRefresh.id,
+        type: TOKEN_TYPE.REFRESH,
+      }).exec();
+      if (!refreshTokenDoc) { throw new Error('invalid-refresh-token'); }
       const user = await UserModel.findById(decodedRefresh.id).exec();
-      const accessToken = await generateAccessToken(user);
+      const accessToken = jwt.sign({
+        id: user._id,
+        email: user.email,
+        type: TOKEN_TYPE.ACCESS,
+        lastName: user.lastName,
+        firstName: user.firstName,
+        sid: refreshTokenDoc.session,
+      }, process.env.JWT_ACCESS_SECRET, { expiresIn: Number(process.env.JWT_ACCESS_LIFETIME || 3600) });
+      const accessTokenDoc = await new TokenModel({
+        user: user._id,
+        type: TOKEN_TYPE.ACCESS,
+        session: refreshTokenDoc.session,
+        expires: new Date(accessToken.exp),
+      }).save();
       return res.status(StatusCodes.OK).json({ accessToken });
     } catch (error) {
       return res.status(StatusCodes.UNAUTHORIZED).json({ message: 'Authorization data is invalid', errorCode: 'unauthorized' });
@@ -211,7 +205,7 @@ export const refreshAccessToken = async (req, res) => {
 
 export const getMe = async (req, res) => {
   try {
-    const user = await UserModel.findById(req.userId).exec();
+    const user = await UserModel.findById(req.authData?.userId).exec();
     return res.status(StatusCodes.OK).json(user);
   } catch (error) {
     return res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({ error });
